pyproject.toml
[project]
name = "prompt-bi"
version = "1.0.0"
description = "Natural-language BI platform powered by Claude AI"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "aiomysql>=0.2.0",

    # Data processing
    "pandas>=2.2.0",
    "duckdb>=0.10.0",
    "openpyxl>=3.1.0",
    "pyarrow>=15.0.0",

    # AI / LLM
    "anthropic>=0.18.0",

    # Google Drive
    "google-api-python-client>=2.116.0",
    "google-auth-oauthlib>=1.2.0",
    "gspread>=6.0.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",

    # Caching & Realtime
    "redis[hiredis]>=5.0.0",
    "python-socketio>=5.11.0",

    # Export
    "playwright>=1.41.0",

    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "slowapi>=0.1.9",
    "loguru>=0.7.0",
    "sqlparse>=0.4.4",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
